Separate sourcemap-concat tests from non-sourcemap-concat
