+++ lint.js
@@ -3,7 +3,7 @@ var lint = require('mocha-eslint');
 var paths = [
   'lib',
   'test/**.js',
-  'test/fixtures/**.js',
+  'test/helpers/**.js',
   'test/strategies/**.js',
   'bench',
   '**.js'
